requirements.txt
# Core ML
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
lightgbm==4.3.0
xgboost==2.0.3
lifelines==0.27.8  # Survival models

# MLOps
mlflow==2.11.0
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1

# Explainability
shap==0.44.1
fairlearn==0.10.0

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# Storage
boto3==1.34.34  # S3/MinIO
minio==7.2.3

# Orchestration
apache-airflow==2.8.1

# Data transformation
dbt-postgres==1.7.6

# Visualization
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.19.0

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.2.3
